hr_system_api/requirements.txt
Flask
Flask-Cors
flask-openapi3[swagger,redoc,rapidoc,rapipdf,scalar,elements]
Flask-SQLAlchemy
nose2
pydantic
SQLAlchemy
SQLAlchemy-Utils
typing_extensions
werkzeug
qrcode[pil]
reportlab
concurrent-log-handler
requests
python-dotenv
pyjwt[crypto]